note about default security group policy
I was bitten by this when creating a new instance: I misread the rules in the default security group and (wrongly) thought they were allowing traffic. I couldn't figure out what the problem was with my setup until I added an allow rule. It seems important to mention here that this will happen to people if they leave the default configuration on.